В форме добавлен метод обновления кол-ва и цены, метод вывода кол-ва и цены в поля при выборе элементаи  метод обновления товара, обновлен метод обновления товара в InvemtoryManager.cs, добавлен еще один тест формы по обновлению товара. Модифицированы тесты класса InventoryManager.cs

+++ InventoryManagement/InventoryManager.cs
@@ -46,7 +46,7 @@ namespace InventoryManagement
             item.Quantity = newQuantity;
             SaveItems();
         }
-        public void UpdateItemPrice(InventoryItem item, int newPrice)
+        public void UpdateItemPrice(InventoryItem item, decimal newPrice)
         {
             if (item == null)
             {
